Add dynamic scheduling for refueling
In normal operation, refueling is controlled by an interrupt (LTN
Fuel). There are rare situations where that interrupt triggers in
inconvenient moments (e.g. between the arrival of the train at the
temporary "pre-stop" and the actual provider or requester). If this
causes problems, this setting can be turned off. LTN will now
dynamically add temporary refuel stops when arriving at a station or
depot with insufficient fuel. Note that this requires a bit more CPU and
might have an affect on FPS/UPS. Using an interrupt is preferred and
dynamic refueling is considered experimental for now.